+++ include/ZBE/core/tools/math/Point.h
@@ -59,7 +59,7 @@ class _POINT {
         return;
       }
 
-      uint i = 0;
+      size_t i = 0;
       for(auto it : l) {
         data[i++] = it;
       }
@@ -78,7 +78,7 @@ class _POINT {
         return (*this);
       }
 
-      uint i = 0;
+      size_t i = 0;
       for(auto it : l) {
         data[i++] = it;
       }
